backend/requirements.txt
fastapi
uvicorn
motor
pydantic
google-generativeai
python-dotenv
httpx
duckduckgo-search
slowapi
pytest